// src/interfaces/standings-interface.ts
// Generated by https://quicktype.io

import { Paging, ParametersInterface, ResponseInterface } from ".";

export interface StandingsInterface {
    get:        string;
    parameters: ParametersInterface;
    errors:     any[];
    results:    number;
    paging:     Paging;
    response:   ResponseInterface[];
}