// src/interfaces/top-scorer-interface.ts
// Generated by https://quicktype.io

import { Paging, ParametersInterface, ResponseInterface } from ".";

export interface TopScorerInterface {
    get:        string;
    parameters: ParametersInterface;
    errors:     any[];
    results:    number;
    paging:     Paging;
    response:   ResponseInterface[];
}